packages/analysis/requirements.txt
yt-dlp
openai-whisper
pandas
argparse
rich 